requirements.txt
Flask==2.2.2
langchain_community==0.2.0
langchain_core==0.2.0
langchain_openai==0.1.7
openai==1.30.1
paddleocr==2.7.3
paddleocr.egg==info
pdf2image==1.17.0
Pillow==10.3.0
PyMuPDF==1.20.2
python-dotenv==1.0.1
python_box==7.1.1
PyYAML==6.0.1
waitress==3.0.0
